+++ pharus/interface.py
@@ -16,42 +16,21 @@ class _DJConnector:
     """Primary connector that communicates with a DataJoint database server."""
 
     @staticmethod
-    def _attempt_login(database_address: str, username: str, password: str):
-        """
-        Attempts to authenticate against database with given username and address.
-
-        :param database_address: Address of database
-        :type database_address: str
-        :param username: Username of user
-        :type username: str
-        :param password: Password of user
-        :type password: str
-        """
-        dj.config["database.host"] = database_address
-        dj.config["database.user"] = username
-        dj.config["database.password"] = password
-
-        # Attempt to connect return true if successful, false is failed
-        dj.conn(reset=True)
-
-    @staticmethod
-    def _list_schemas(jwt_payload: dict) -> list:
+    def _list_schemas(connection: dj.Connection) -> list:
         """
         List all schemas under the database.
 
-        :param jwt_payload: Dictionary containing databaseAddress, username, and password
-            strings
-        :type jwt_payload: dict
+        :param connection: User's DataJoint connection object
+        :type connection: dj.Connection
         :return: List of schemas names in alphabetical order (excludes ``information_schema``,
             ``sys``, ``performance_schema``, ``mysql``)
         :rtype: list
         """
-        _DJConnector._set_datajoint_config(jwt_payload)
 
         # Attempt to connect return true if successful, false is failed
         return [
             row[0]
-            for row in dj.conn().query(
+            for row in connection.query(
                 """
                 SELECT SCHEMA_NAME FROM information_schema.schemata
                 WHERE SCHEMA_NAME NOT IN (
@@ -63,23 +42,26 @@ class _DJConnector:
         ]
 
     @staticmethod
-    def _list_tables(jwt_payload: dict, schema_name: str) -> dict:
+    def _list_tables(
+        connection: dj.Connection,
+        schema_name: str,
+    ) -> dict:
         """
         List all tables and their type given a schema.
 
-        :param jwt_payload: Dictionary containing databaseAddress, username, and password
-            strings
-        :type jwt_payload: dict
+        :param connection: User's DataJoint connection object
+        :type connection: dj.Connection
         :param schema_name: Name of schema to list all tables from
         :type schema_name: str
         :return: Contains a key for each table type where values are the respective list of
             table names
         :rtype: dict
         """
-        _DJConnector._set_datajoint_config(jwt_payload)
 
         # Get list of tables names
-        tables_name = dj.Schema(schema_name, create_schema=False).list_tables()
+        tables_name = dj.Schema(
+            schema_name, create_schema=False, connection=connection
+        ).list_tables()
         # Dict to store list of table name for each type
         tables_dict_list = dict(manual=[], lookup=[], computed=[], imported=[], part=[])
         # Loop through each table name to figure out what type it is and add them to
@@ -283,14 +265,15 @@ class _DJConnector:
 
     @staticmethod
     def _get_table_definition(
-        jwt_payload: dict, schema_name: str, table_name: str
+        connection: dj.Connection,
+        schema_name: str,
+        table_name: str,
     ) -> str:
         """
         Get the table definition.
 
-        :param jwt_payload: Dictionary containing databaseAddress, username, and password
-            strings
-        :type jwt_payload: dict
+        :param connection: User's DataJoint connection object
+        :type connection: dj.Connection
         :param schema_name: Name of schema
         :type schema_name: str
         :param table_name: Table name under the given schema; must be in camel case
@@ -298,22 +281,25 @@ class _DJConnector:
         :return: Definition of the table
         :rtype: str
         """
-        _DJConnector._set_datajoint_config(jwt_payload)
 
         local_values = locals()
-        local_values[schema_name] = dj.VirtualModule(schema_name, schema_name)
+        local_values[schema_name] = dj.VirtualModule(
+            schema_name, schema_name, connection=connection
+        )
         return getattr(local_values[schema_name], table_name).describe()
 
     @staticmethod
     def _insert_tuple(
-        jwt_payload: dict, schema_name: str, table_name: str, tuple_to_insert: dict
+        connection: dj.Connection,
+        schema_name: str,
+        table_name: str,
+        tuple_to_insert: dict,
     ):
         """
         Insert record as tuple into table.
 
-        :param jwt_payload: Dictionary containing databaseAddress, username, and password
-            strings
-        :type jwt_payload: dict
+        :param connection: User's DataJoint connection object
+        :type connection: dj.Connection
         :param schema_name: Name of schema
         :type schema_name: str
         :param table_name: Table name under the given schema; must be in camel case
@@ -321,22 +307,25 @@ class _DJConnector:
         :param tuple_to_insert: Record to be inserted
         :type tuple_to_insert: dict
         """
-        _DJConnector._set_datajoint_config(jwt_payload)
 
-        schema_virtual_module = dj.VirtualModule(schema_name, schema_name)
+        schema_virtual_module = dj.VirtualModule(
+            schema_name, schema_name, connection=connection
+        )
         getattr(schema_virtual_module, table_name).insert(tuple_to_insert)
 
     @staticmethod
     def _record_dependency(
-        jwt_payload: dict, schema_name: str, table_name: str, restriction: list = []
+        connection: dj.Connection,
+        schema_name: str,
+        table_name: str,
+        restriction: list = [],
     ) -> list:
         """
         Return summary of dependencies associated with a restricted table. Will only show
         dependencies that user has access to.
 
-        :param jwt_payload: Dictionary containing databaseAddress, username, and password
-            strings
-        :type jwt_payload: dict
+        :param connection: User's DataJoint connection object
+        :type connection: dj.Connection
         :param schema_name: Name of schema
         :type schema_name: str
         :param table_name: Table name under the given schema; must be in camel case
@@ -347,8 +336,10 @@ class _DJConnector:
         :return: Tables that are dependent on specific records.
         :rtype: list
         """
-        _DJConnector._set_datajoint_config(jwt_payload)
-        virtual_module = dj.VirtualModule(schema_name, schema_name)
+
+        virtual_module = dj.VirtualModule(
+            schema_name, schema_name, connection=connection
+        )
         table = getattr(virtual_module, table_name)
         attributes = table.heading.attributes
         # Retrieve dependencies of related to retricted
@@ -379,14 +370,16 @@ class _DJConnector:
 
     @staticmethod
     def _update_tuple(
-        jwt_payload: dict, schema_name: str, table_name: str, tuple_to_update: dict
+        connection: dj.Connection,
+        schema_name: str,
+        table_name: str,
+        tuple_to_update: dict,
     ):
         """
         Update record as tuple into table.
 
-        :param jwt_payload: Dictionary containing databaseAddress, username, and password
-            strings
-        :type jwt_payload: dict
+        :param connection: User's DataJoint connection object
+        :type connection: dj.Connection
         :param schema_name: Name of schema
         :type schema_name: str
         :param table_name: Table name under the given schema; must be in camel case
@@ -394,10 +387,11 @@ class _DJConnector:
         :param tuple_to_update: Record to be updated
         :type tuple_to_update: dict
         """
-        conn = _DJConnector._set_datajoint_config(jwt_payload)
 
-        schema_virtual_module = dj.VirtualModule(schema_name, schema_name)
-        with conn.transaction:
+        schema_virtual_module = dj.VirtualModule(
+            schema_name, schema_name, connection=connection
+        )
+        with connection.transaction:
             [
                 getattr(schema_virtual_module, table_name).update1(t)
                 for t in tuple_to_update
@@ -405,7 +399,7 @@ class _DJConnector:
 
     @staticmethod
     def _delete_records(
-        jwt_payload: dict,
+        connection: dj.Connection,
         schema_name: str,
         table_name: str,
         restriction: list = [],
@@ -414,9 +408,8 @@ class _DJConnector:
         """
         Delete a specific record based on the restriction given.
 
-        :param jwt_payload: Dictionary containing databaseAddress, username, and password
-            strings
-        :type jwt_payload: dict
+        :param connection: User's DataJoint connection object
+        :type connection: dj.Connection
         :param schema_name: Name of schema
         :type schema_name: str
         :param table_name: Table name under the given schema; must be in camel case
@@ -427,9 +420,10 @@ class _DJConnector:
         :param cascade: Allow for cascading delete, defaults to ``False``
         :type cascade: bool, optional
         """
-        _DJConnector._set_datajoint_config(jwt_payload)
 
-        schema_virtual_module = dj.VirtualModule(schema_name, schema_name)
+        schema_virtual_module = dj.VirtualModule(
+            schema_name, schema_name, connection=connection
+        )
 
         # Get table object from name
         table = _DJConnector._get_table_object(schema_virtual_module, table_name)
@@ -507,19 +501,3 @@ class _DJConnector:
                 else attribute_filter["value"]
             )
         return f"{attribute_filter['attributeName']}{operation}{value}"
-
-    @staticmethod
-    def _set_datajoint_config(jwt_payload: dict) -> dj.connection.Connection:
-        """
-        Method to set credentials for database.
-
-        :param jwt_payload: Dictionary containing databaseAddress, username, and password
-            strings
-        :type jwt_payload: dict
-        :return: DataJoint connection object.
-        :rtype: :class:`~datajoint.connection.Connection`
-        """
-        dj.config["database.host"] = jwt_payload["databaseAddress"]
-        dj.config["database.user"] = jwt_payload["username"]
-        dj.config["database.password"] = jwt_payload["password"]
-        return dj.conn(reset=True)